Add soc sw texture mapping

+++ soc/src/examples/3drenderer/src/display.c
@@ -53,7 +53,7 @@ void draw_grid(void) {
             color_buffer[(window_width * y) + x] = 0xF333;
 }
 
-void draw_pixel(int x, int y, uint32_t color) {
+void draw_pixel(int x, int y, uint16_t color) {
     if (x >=0 && x < window_width && y >= 0 && y < window_height)
         color_buffer[(window_width * y) + x] = color;
 }
