+++ soc/src/examples/3drenderer/src/texture.h
@@ -13,6 +13,6 @@ extern int texture_height;
 
 extern const uint8_t REDBRICK_TEXTURE[];
 
-extern uint16_t* mesh_texture;
+extern uint32_t* mesh_texture;
 
 #endif
